Dealing with large dataset
The number of cells I'm trying to look at is increasing past the point
that can be loaded into ram. I am working to save analyzed data to .csv
and just running analyses on the new cells. It's a work in progress, and
may have to change as I decide to add more analyses. Ideally, I can use
the p.csv since it has all the data, as long as I don't want to redo
spike density functions or saccade identification.