Merged vergence project
I started a new folder for SOA and NRTP analysis. There is a new R
project and a new Rmd file. The data file associated with this is
NRTPt.RDS. The .csv files I have saved on my desktop. I imagine that
this could be modified to load directly from the .mat files in the
future, but that would be a very slow operation I believe.