modeling and enhancement algorithm
I made a function to detect vergence enhancement. I think the algorithm
may be general enough to detect any event using a two threshold
criteria. First, it marks periods where the velocity is above the lower
threshold. Then it rejects any where the velocity fails to rise above
the higher threshold.

This has not yet been tested for accuracy. Next step is to incorporate
it into the modeling. It may be as simple as using this function to mark
transients for the model fitting.