OMN INC and more
I've been working on the INC and OMN analysis. I made a few changes to
the Matlab spike sorting GUI. I added the cluster number to the file
name and I added a button that automatically clicks yes on all the
options. It knows that Patos and Pilchuck are exotropes, but will need
to be updated for others. I thought it might be better to just add
toggle buttons on the GUI rather than having the yes/no popups.