INC stim analysis updated OMN analysis added
I am basically using the same code to analyze fixations for OMN and INC
but I want to keep separate notebooks so I can handle plotting
differently. I created "fixationalysisfunctions.R" to hold the functions
that are used by both INC and OMN but It's a work in progress.

Fixed a bug in savecsv.m that wasn't identifying DC as an exotrope when
quicksave was used. This function will need to be manually updated with
a list of exotropes